Updated CEFX validators to new VirSat API- (Task #113)
