+++ de.dlr.sc.virsat.model.extension.cefx/plugin.xml
@@ -2,17 +2,17 @@
 <?eclipse version="3.4"?>
 <plugin>
 	<extension point="de.dlr.sc.virsat.model.Concept">
-	<concept
-	          id="de.dlr.sc.virsat.model.extension.cefx"
-	          version="1.0"
-	          xmi="concept/concept.xmi">
-	    </concept>
+		<concept
+			id="de.dlr.sc.virsat.model.extension.cefx"
+			version="1.0"
+			xmi="concept/concept.xmi">
+		</concept>
 	</extension>
 	<extension point="de.dlr.sc.virsat.model.DvlmValidator">
 		<dvlmValidator>
 			<seiValidator
 				id="de.dlr.sc.virsat.model.extension.cefx"
-				class="de.dlr.sc.virsat.model.extension.cefx.validator.StructuralElementInstanceValidator">
+				class="de.dlr.sc.virsat.model.extension.cefx.validator.CefxValidator">
 			</seiValidator>
 		</dvlmValidator>
 	</extension>
